requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.2
python-dotenv==1.0.0
pydantic[email]==2.5.3
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.6
qrcode[pil]==7.4.2
Pillow==10.2.0
boto3==1.34.34
requests==2.31.0
